refactor(ipc): split commands/mod.rs into peer modules (#431 item 1)
Pure code-move refactor. The 3,329-LOC `commands/mod.rs` had grown
past the point where the contributor map at the top of the file
substituted for actual module boundaries. Lifts five well-bounded
command groups into peer files following the existing `meeting.rs`
/ `models.rs` / `dictionary.rs` / `macos.rs` / `export.rs` pattern:

- `commands/history.rs` (252 LOC) — list/search/export/delete/count/
  clear + dictation stats. Tests + the pure CSV helper move with it.
- `commands/settings.rs` (276 LOC) — get/set for hud / sound_cues /
  sound_cue_start / sound_cue_complete / diarization /
  inference_threads / meeting_autostart_mode. The `_inner` variants
  stay `pub(crate)` so the boot path + tests can call them without
  Tauri.
- `commands/system.rs` (197 LOC) — show_main_window, first-run
  flag trio (get / mark / reset), autostart-path-stale pair,
  check_for_updates pair.
- `commands/ptt.rs` (155 LOC) — PttConfig + ptt_get_config +
  ptt_set_config.
- `commands/diarizer.rs` (385 LOC) — DiarizeModelStatus +
  download/remove + the in-process `swap_diarizer_after_download`
  helper.

`commands/mod.rs` drops to 2,215 LOC, retaining the dictation hot
path (start_dictation / stop_dictation + helpers), shared types
(IpcError / IpcResult / DictationResult), and the dictation-side
test surface. The dictation pipeline is tightly coupled enough that
moving it would have outsized regression risk for cosmetic LOC win.

Each `tauri::generate_handler!` registration in `src-tauri/src/lib.rs`
gains a fully-qualified path (`ipc::commands::history::history_list`
etc.) per the four-place sync rule + `learnings.md` 2026-04-25 — `pub
use` re-exports do not carry the macro's hidden `__cmd__<name>` symbol.

372 lib tests pass, clippy `-D warnings` clean, npm run check clean,
87 Playwright specs pass.
