+++ src-tauri/src/lib.rs
@@ -546,17 +546,17 @@ pub fn run() {
         })
         .invoke_handler(tauri::generate_handler![
             ipc::commands::audio_list_sources,
-            ipc::commands::show_main_window,
+            ipc::commands::system::show_main_window,
             ipc::commands::start_dictation,
             ipc::commands::stop_dictation,
-            ipc::commands::history_list,
-            ipc::commands::history_search,
-            ipc::commands::history_export_row_csv,
+            ipc::commands::history::history_list,
+            ipc::commands::history::history_search,
+            ipc::commands::history::history_export_row_csv,
             ipc::commands::export::history_export_bundle,
-            ipc::commands::history_delete,
-            ipc::commands::history_count,
-            ipc::commands::history_clear,
-            ipc::commands::get_dictation_stats,
+            ipc::commands::history::history_delete,
+            ipc::commands::history::history_count,
+            ipc::commands::history::history_clear,
+            ipc::commands::history::get_dictation_stats,
             ipc::commands::dictionary::replacements_list,
             ipc::commands::dictionary::replacement_create,
             ipc::commands::dictionary::replacement_update,
@@ -570,31 +570,31 @@ pub fn run() {
             ipc::commands::models::model_download,
             ipc::commands::models::model_cancel_download,
             ipc::commands::models::model_remove,
-            ipc::commands::get_first_run_completed,
-            ipc::commands::mark_first_run_completed,
-            ipc::commands::reset_first_run,
-            ipc::commands::get_hud_enabled,
-            ipc::commands::set_hud_enabled,
-            ipc::commands::get_sound_cues_enabled,
-            ipc::commands::set_sound_cues_enabled,
-            ipc::commands::get_sound_cue_start_enabled,
-            ipc::commands::set_sound_cue_start_enabled,
-            ipc::commands::get_sound_cue_complete_enabled,
-            ipc::commands::set_sound_cue_complete_enabled,
-            ipc::commands::get_meeting_autostart_mode,
-            ipc::commands::set_meeting_autostart_mode,
-            ipc::commands::get_diarization_enabled,
-            ipc::commands::set_diarization_enabled,
-            ipc::commands::get_inference_threads,
-            ipc::commands::set_inference_threads,
-            ipc::commands::get_diarizer_model_status,
-            ipc::commands::download_diarizer_model,
-            ipc::commands::remove_diarizer_model,
-            ipc::commands::get_autostart_path_status,
-            ipc::commands::retry_autostart_registration,
-            ipc::commands::check_for_updates,
-            ipc::commands::ptt_get_config,
-            ipc::commands::ptt_set_config,
+            ipc::commands::system::get_first_run_completed,
+            ipc::commands::system::mark_first_run_completed,
+            ipc::commands::system::reset_first_run,
+            ipc::commands::settings::get_hud_enabled,
+            ipc::commands::settings::set_hud_enabled,
+            ipc::commands::settings::get_sound_cues_enabled,
+            ipc::commands::settings::set_sound_cues_enabled,
+            ipc::commands::settings::get_sound_cue_start_enabled,
+            ipc::commands::settings::set_sound_cue_start_enabled,
+            ipc::commands::settings::get_sound_cue_complete_enabled,
+            ipc::commands::settings::set_sound_cue_complete_enabled,
+            ipc::commands::settings::get_meeting_autostart_mode,
+            ipc::commands::settings::set_meeting_autostart_mode,
+            ipc::commands::settings::get_diarization_enabled,
+            ipc::commands::settings::set_diarization_enabled,
+            ipc::commands::settings::get_inference_threads,
+            ipc::commands::settings::set_inference_threads,
+            ipc::commands::diarizer::get_diarizer_model_status,
+            ipc::commands::diarizer::download_diarizer_model,
+            ipc::commands::diarizer::remove_diarizer_model,
+            ipc::commands::system::get_autostart_path_status,
+            ipc::commands::system::retry_autostart_registration,
+            ipc::commands::system::check_for_updates,
+            ipc::commands::ptt::ptt_get_config,
+            ipc::commands::ptt::ptt_set_config,
             ipc::commands::macos::open_macos_privacy_pane,
             ipc::commands::macos::diagnose_macos_permissions,
             ipc::commands::macos::reset_macos_permissions,
